Version: 0.13.33-alpha.5 Chat app refactored for Llama 3.1 8B FP4 default; VRAM fixes, quantization updates
Chat App:
- Chat app nw sets Llama 3.1 8B FP4 as the default model (was 3.2 3B BF16)
- Model switching now destroys old model before loading new to avoid VRAM overcommit and shared memory spill
- Fix Llama 8B loader to use correct "llama31" family filename

Mila:
- CudaLinearOp now fetches shared BF16 staging buffer at forward time, preventing stale pointer bugs and reducing memory fragmentation
- LlamaTransformer execution context reordered for safe CUDA stream synchronization on destruction
- Update docs and roadmap for new defaults, bug fixes, and quantization validation

+++ Mila/Tools/Converters/Llama/convert_weights.py
@@ -157,7 +157,7 @@ def convert_llama( model_name: str, output_path: str, dtype: str = 'float32' ):
     print( f"Loading {model_name} from HuggingFace (dtype={dtype})..." )
 
     try:
-        model = AutoModelForCausalLM.from_pretrained( model_name, torch_dtype=torch_dtype )
+        model = AutoModelForCausalLM.from_pretrained( model_name, dtype=torch_dtype )
     except Exception as e:
         _check_hf_error( model_name, e )
